requirements.txt
Flask
Flask-Session
gunicorn
requests
numpy
Werkzeug
tensorflow-cpu
scikit-image
scikit-learn
opencv-python